payment process completed

+++ products/models.py
@@ -29,7 +29,7 @@ class Product(models.Model):
     title = models.CharField(max_length=200)
     description = RichTextField()
     brand = models.CharField(max_length=200, blank=True)
-    price = models.DecimalField(decimal_places=2, max_digits=6)
+    price = models.PositiveIntegerField()
     datetime_created = models.DateTimeField(auto_now_add=True)
     datetime_modified = models.DateTimeField(auto_now=True)
     active = models.BooleanField(default=True)
